requirements.txt
Flask==2.0.1
celery==5.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
Flask-Mail==0.9.1
googletrans==3.1.0a0
python-dateutil==2.8.2
reportlab==3.6.1
python-dotenv==0.19.0
tenacity==9.0.0
